refactor(coquille): les 35 écouteurs quittent l'amorce — il ne reste qu'une séquence
`amorce.js` passe de 501 à 187 lignes, et ne porte plus qu'UN `addEventListener` : celui du service
worker. Le câblage tient en sept appels.

Quatre modules naissent, chacun avec une raison qui tient en une phrase :

— **`controles.js`** — la barre de filtres et les réglages de chaque vue. Trois règles y sont
  écrites parce qu'aucune n'est devinable : saisie libre = débouncé / menu = immédiat ;
  `rafraichirDifferee` est UNE instance partagée (six timers indépendants rendraient six fois) mais
  `#origin` et `#destTerminal` ont chacun le leur (les fusionner ferait qu'une frappe dans l'un
  annule l'autre) ; `change` et `input` ne sont pas interchangeables sur une case.
— **`corrections-gestes.js`** — la délégation sur `#corrections`, parent des trois portails de la
  vue, et l'ordre de ses branches : `.al-del` en premier (son ✕ porte aussi `.corr-del`),
  `#exportCorrections` avant `#resetAll`.
— **`soute-gestes.js`** — les trois cartes du bandeau.
— **`listes.js`** — les trois `<datalist>` et leur drapeau. L'état y est PRIVÉ derrière
  `listesPretes()` : exporter la liaison `let` marcherait, mais un importateur qui la recopie dans
  une `const` la figerait à `false` et reconstruirait 114 + 92 `<option>` à chaque focus.

Et trois modules existants reçoivent les délégations qui appellent leurs propres gestes :
`manifeste-gestes.js` (`#manifest`), `voyage-gestes.js` (les huit du compagnon, dont six sur
`document`), `presse-papiers.js` (`#planCopy` et `#share`).

Deux corps de geste montent au passage dans `corrections-actions.ts` — `effacerStation()` et
`revenirAUEX()` — parce qu'ils étaient écrits DANS un écouteur : ils y gèlent les jambes avant
d'effacer un volume (#48), et cette règle mérite mieux qu'une branche de `switch`.

Aucune délégation ne devient un `onClick`. Elles sont posées sur des conteneurs que les portails
remplissent sans les posséder : elles traversent les rendus (ADR-012 §2), et les convertir
doublerait `pinLegsForVolume`, `addLegLine` et `addSuggestion`, qui ne sont pas idempotents.

Vérification :
```
node --test          → tests 490 | pass 490 | fail 0
npx playwright test  → 246 tests | 244 passed | 2 skipped | 0 failed (1,6 min)
npx tsc --noEmit     → aucune sortie
npm run build:site   → ✓ built | précache 20 entrées
```

Trois identifiants LIBRES ont été attrapés avant les tests, par un contrôle explicite « chaque
`brancher*` appelé est-il importé ? » : `brancherPressePapiers`, `brancherGestesManifeste` et
`brancherGestesVoyage`. `tsc` ne lit pas ce fichier, `vite build` non plus — c'est la même classe de
défaut qui a coûté deux régressions au lot précédent, et elle se cherche à la main.
